requirements.txt
Flask==2.3.3
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
werkzeug==2.3.7
gunicorn==21.2.0
google-generativeai==0.3.0
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.4 